third_party/blink/web_tests/svg/dynamic-updates/SVGFEColorMatrixElement-dom-type-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'type' attribute of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS matrixElement.getAttribute('type') is "saturate"
PASS saturateElement.getAttribute('type') is "hueRotate"
PASS hueRotateElement.getAttribute('type') is "luminanceToAlpha"
PASS luminanceToAlphaElement.getAttribute('type') is "matrix"
PASS matrixElement.getAttribute('type') is "matrix"
PASS saturateElement.getAttribute('type') is "saturate"
PASS hueRotateElement.getAttribute('type') is "hueRotate"
PASS luminanceToAlphaElement.getAttribute('type') is "luminanceToAlpha"
PASS successfullyParsed is true

TEST COMPLETE
